Move populate_fresh_nodes and nearby_nodes to common
